Finished first working version of multivariate ESS
