+++ examples/multi_ess.py
@@ -4,181 +4,11 @@
 
 import numpy as np
 
-# from kanga.stats import multi_ess
-
-# mport statsmodels.api as sm
-# from statsmodels.tsa.ar_model import AutoReg
-from statsmodels.tsa.ar_model import ar_select_order
-from statsmodels.tsa.stattools import acovf
-
-# https://stats.stackexchange.com/questions/371792/sum-of-autocovariances-for-arp-model/372006#372006
-# https://arxiv.org/pdf/1804.05975.pdf
-# http://stat.wharton.upenn.edu/~steele/Courses/956/Resource/YWSourceFiles/WhyNotToUseYW.pdf
-# https://ieeexplore.ieee.org/document/4547058
-# https://arxiv.org/pdf/0811.1729.pdf
-# https://arxiv.org/pdf/1809.04541.pdf
-
-# %%
-
-def approx_via_arp(x):
-    n = len(x)
-
-    ar_selection = ar_select_order(
-        x,
-        min(int(np.floor(n)), int(np.floor(10 * np.log10(n)))),
-        ic='aic',
-        trend='n',
-        seasonal=False
-    )
-
-    ar_selection_order = ar_selection.ar_lags[-1]
-
-    gammas = acovf(x, nlag=ar_selection_order-1, fft=False)
-
-    model_fit = ar_selection.model.fit()
-    
-    param_sum = model_fit.params.sum()
-
-    asympt_var = model_fit.sigma2 / ((1 - param_sum) ** 2)
-
-    if (ar_selection_order != 0):
-        Gamma = 0
-
-        for i in range(1, ar_selection_order+1):
-            for k in range(1, i+1):
-                Gamma = Gamma + model_fit.params[i-1] * k * gammas[i-k]
-
-        Gamma = 2 * (
-            Gamma + 0.5 * (asympt_var - gammas[0]) * (model_fit.params * range(1, ar_selection_order + 1)).sum()
-            ) / (1 - param_sum)
-    else:
-        Gamma = 0
-
-    return Gamma, asympt_var
-
-# %%
-
-def optimal_batch_size(x):
-    num_iters = x.shape[0]
-
-    ar_approximation = np.apply_along_axis(approx_via_arp, 0, x) ** 2
-    coeff = (ar_approximation[0, ].sum() / ar_approximation[1, ].sum()) ** (1 / 3)
-
-    b = int(np.floor((num_iters ** (1 / 3)) * coeff))
-    if b == 0:
-        b = 1
-
-    return b
-
-# %%
-
-# Definition of batch means: see p. 7, eq. 6 in https://arxiv.org/pdf/0811.1729.pdf
-
-def batch_mean(x, b):
-    return [np.mean(x[i*b:((i+1)*b)]) for i in range(len(x) // b)]
-
-def batch_var(x, b):
-    return np.var(batch_mean(x, b), ddof=1)
-
-# %%
-
-def bm_cov(x, b):
-    col_batch_means = np.apply_along_axis(lambda v: batch_mean(v, b), 0, x)
-    return b * np.cov(col_batch_means.transpose())
-
-# %%
-
-# b is the batch size
-def mc_cov(x, b=None, r=3):
-    if b is None:
-        b = optimal_batch_size(x)
-
-    if ((r < 0) or (r > 5)):
-        raise ValueError(' r={}. Pick a value of r in between 1 and 5'.format(r))
-
-    if (b == 1 and r != 1):
-        raise ValueError('r={} and b={}. If b=1, then r should be set to 1 too'.format(r, b))
-
-    br = int(np.floor(b / r))
-    if (br < 1):
-        raise ValueError('floor(b/r)={}<1. Increase b or decrease r'.format(br))
-
-    cov_mat = bm_cov(x, b)
-
-    # Possibly use Lugsail lag window (see p. 4, eq. 2 in https://arxiv.org/pdf/1809.04541.pdf)
-    if (r > 1):
-        lug_cov_mat = 2 * cov_mat - bm_cov(x, br)
-
-        if (lug_cov_mat.diagonal().prod() > 0):
-            cov_mat = lug_cov_mat
-
-    return cov_mat
-
-# %%
-
-def multi_ess(x, b=None, r=3):
-    num_iters, num_pars = x.shape
-
-    cov_mat = np.cov(x.transpose())
-
-    cov_mat_det = np.linalg.det(cov_mat)
-    mc_cov_mat_det = np.linalg.det(mc_cov(x, b=b, r=r))
-
-    return num_iters * ((cov_mat_det / mc_cov_mat_det) ** (1/num_pars))
+from kanga.stats import multi_ess
 
 # %% Read chains
 
 chains = np.genfromtxt('chain01.csv', delimiter=',')
-
-# %%
-
-Gamma, asympt_var = approx_via_arp(chains[:, 0])
-
-print(Gamma)
-print(asympt_var)
-
-# %%
-
-print(optimal_batch_size(chains))
-
-# %%
-
-from statsmodels.regression.linear_model import yule_walker
-
-# %%
-
-a1, b1 = yule_walker(chains[:, 0], order=2, method='mle')
-print(a1)
-print(b1**2)
-
-# %%
-
-from statsmodels.regression.linear_model import burg
-
-# %%
-
-a2, b2 = burg(chains[:, 0], order=2)
-print(a2)
-print(b2)
-
-# %%
-
-num_iters, num_pars = chains.shape
-
-selection = ar_select_order(
-    chains[:, 0],
-    min(int(np.floor(num_iters)), int(np.floor(10 * np.log10(num_iters)))),
-    ic='aic',
-    trend='n',
-    seasonal=False
-)
-print(selection.ar_lags)
-res = selection.model.fit()
-# print(res.summary())
-print(res.params)
-print(res.sigma2)
-
-# %%
 
 # %% Compute multivariate ESS
