+++ kanga/stats/__init__.py
@@ -1,4 +1,8 @@
+from .batch_mean import batch_mean
+from .batch_var import batch_var
+from .bm_cov import bm_cov
 from .cor import cor
+from .mc_cov import mc_cov
 from .mean_cor import mean_cor
 from .multi_ess import multi_ess
 from .rhat import rhat
